Initial Prototype with Attacking an Enemy
Rearranged a lot of files to make more sense.  The player is able to
attack now with the J button, and the attack range is indicated by a
flashing white box.  The player will also get hurt if touching the
enemy, and will fly back and be slightly immobile for a couple of
seconds.