Mucking about with jump and level design
I created a basic map to test jumping versus walking under platforms.
Before, the sprite could either jump onto a platform or walk under it,
but it couldn't do both.  So I adjusted its jump height to enable both.
Modified TestLevel so that it runs using this test map, but that's not
too important so I'll let you guys do that yourselves