Made a flat corridor
The look of this one is closer to the look of Michael Gelon's prototype.
The corridor itself is probably too long, so tweaking shall be
necessary.  Made two levels that utilize it - one for introducing
movement and jumping (pending the implementation of the crate class),
one for introducing enemy interaction.