Added the Informant and Bug Fixes
Made the boxes have a title bar and can be automatically closed.  Fixed
a bug in FlxSave to ensure that saves across different websites work.
Fixed a bug in FlxCamera that made zoomed flash screens not have the
camera scroll correctly.  Fixed a bug in FlxTimer that added a timer
every time after it was paused, making it go twice as fast as
previously.