Added Sounds and Periodic Sounds
Added a message box and running sound.  Added in periodic sounds, which
are sounds that periodically play.  You can also associate a periodic
sound with a state, so that that sound will only play during that state.
Added in an identifier to the game, so everytime the game is played, a
random number is generated, which is then used to help identify which
replay is being saved to the server.